requirements.txt
requests==2.31.0
pandas==2.1.0
openai==1.3.0
chromadb==0.4.15
python-dotenv==1.0.0
tqdm==4.66.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
beautifulsoup4==4.12.2 